added moving to commons the schema of workflows + policies
