+++ packages/common/src/index.ts
@@ -11,3 +11,13 @@ export {
   sleep,
   uniqueArray,
 } from './utils';
+
+export {
+  getDocumentId,
+  type DefaultContextSchema,
+  type TDefaultSchemaDocumentPage,
+  ghanaDocuments,
+  certificateOfResidenceGH,
+  getDocumentsByCountry,
+  type TDocument,
+} from './schemas';
