+++ packages/common/src/schemas/documents/default-context-schema.json
