+++ packages/common/src/schemas/documents/workflow/documents/schemas/GH.ts
@@ -1,6 +1,6 @@
-import { Document } from '../types';
+import { TDocument } from '../types';
 
-export const certificateOfResidenceGH: Document = {
+export const certificateOfResidenceGH: TDocument = {
   category: 'proof_of_address',
   type: 'water_bill',
   issuer: {
@@ -38,7 +38,7 @@ export const certificateOfResidenceGH: Document = {
   },
 };
 
-export const ghanaDocuments: Document[] = [
+export const ghanaDocuments: TDocument[] = [
   {
     category: 'proof_of_address',
     type: 'water_bill',
