+++ packages/common/src/schemas/documents/workflow/documents/types.ts
@@ -1,0 +1,5 @@
+import { DefaultContextSchema } from '@/workflow/schemas/context';
+
+export type TDocument = Omit<DefaultContextSchema['documents'][number], 'pages' | 'properties'> & {
+  propertiesSchema: any;
+};
