--- services/workflows-service/src/documents/types.ts
@@ -1,5 +1,0 @@
-import { DefaultContextSchema } from '@/workflow/schemas/context';
-
-export type Document = Omit<DefaultContextSchema['documents'][number], 'pages' | 'properties'> & {
-  propertiesSchema: any;
-};
